Add files via upload
Hi Team, 

Here is my solution , to the assignment which I was asked to complete. I tried to deduce the solution upto best of my ability. Excuse me for my mistakes. 
I was asked to rederive the list of order page wise . I was not able to do so. 
Please go through the code and please tell me my mistakes. 

Thanks and Regards,
Padmaj Verma.